requirements.txt
absl-py==2.1.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.8.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
binance==0.3
bitarray==2.8.1
blinker==1.9.0
boto3==1.37.4
botocore==1.37.4
cbor2==5.4.6
certifi
charset-normalizer==3.2.0
click==8.1.8
cloudpickle==3.1.1
contourpy==1.2.1
crytic-compile==0.3.4
cycler==0.12.1
Cython==3.0.11
cytoolz==0.12.2
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
Farama-Notifications==0.0.4
filelock==3.17.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.0
fsspec==2025.2.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.66.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h5py==3.11.0
hexbytes==0.3.1
httpcore==1.0.7
httpx==0.28.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lru-dict==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
maturin==1.2.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.4
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==23.1
pandas==2.2.2
parsimonious==0.9.0
pillow==10.4.0
pluggy==1.5.0
prettytable==3.8.0
protobuf==4.24.1
pycryptodome==3.18.0
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2024.1
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.7.1
rlp==3.0.0
rpds-py==0.9.2
s3transfer==0.11.3
schedule==1.2.2
scikeras==0.13.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
slither-analyzer==0.9.6
sniffio==1.3.1
solc-select==1.0.4
stable_baselines3==2.5.0
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==2.4.0
threadpoolctl==3.5.0
toml
toolz==0.12.0
torch==2.6.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.4
wcwidth==0.2.6
web3==6.9.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.9.2
